api/app/requirements.txt
fastapi[standard]==0.112.0
uvicorn==0.32.0
sqlmodel==0.0.22
psycopg2==2.9.10
bcrypt==4.2.0
PyJWT==2.9.0
types-PyJWT==1.7.1
jsonpickle==3.4.2
